+++ changelog.md
@@ -1,2 +1,2 @@
 ### Benched 1.2.2
-- Initial release of Benched for Minecraft 1.21.5
+- Initial release of Benched for Minecraft 1.21.6